requirements.txt.txt
torch
torchvision 
torchaudio
ultralytics
tensorboard
pytorch-ignite
